requirements.txt
streamlit
pymupdf
python-dotenv
apify-client
groq
